requirements.txt
markdown>=3.5
pyyaml>=6.0
requests>=2.28
